execute tests, publish to mvn local
test scripts are not committed executable at the moment, so
execute them explicit with bash. they search in maven local
for the plugin, enable publish to maven local again.

fix the test build, in some cases it looks upwards for a
settings file, and errors out. put an empty settings.gradle
so it is clear its an independent gradle build.